# swagger.yaml
swagger: '2.0'
info:
  title: College Management System API Documentation
  version: v1
basePath: /api
...
paths:
...
  /tutorials/addNewPage:
    post:
      summary: create new tutorial
      operationId: addNewPage
      tags:
        - tutorial
      parameters:
        - in: body
          name: newPage
          required: true
          schema:
            type: object
            required:
              - title
              - data
            properties:
              title:
                type: string
              data:
                type: string
      responses:
        '201':
          description: Create new page
        '403':
          description: User not authenticated or doesnt have enough priviliges
  /admin/addUser:
    post:
      summary: create new user
      operationId: addUser
      tags:
        - admin
      parameters:
        - in: body
          name: newPage
          required: true
          schema:
            type: object
            required:
              - name
              - email
              - password
              - role
            properties:
              name:
                type: string
              email:
                type: string
              password:
                type: string
              role:
                type: string
      responses:
        '201':
          description: Create new role
        '403':
          description: User not authenticated or doesnt have enough priviliges
  /admin/listUsers:
    get:
      summary: list paginated users
      operationId: listUsers
      tags:
        - admin
      responses:
        '200':
          description: List paginated users
        '403':
          description: User not authenticated or doesnt have enough priviliges
...
  /admin/hostel/allocateRoom:
    post:
      summary: add new room data
      operationId: allocateRoom
      tags:
        - admin
      parameters:
        - in: body
          name: allocateRoom
          required: true
          schema:
            type: object
            required:
              - userid
            properties:
              userid:
                type: string
      responses:
        '200':
          description: list paginated Hostel Requests
        '403':
          description: User not authenticated or doesnt have enough priviliges
...